UIDS-81 Implement review comments
