+++ src/CheckboxButton/index.js
@@ -1,0 +1,3 @@
+import CheckboxButton from './CheckboxButton';
+
+export default CheckboxButton;
