--- src/CheckboxButtonGroup/CheckboxButton.scss
@@ -1,21 +1,0 @@
-@import '../../scss/theme';
-
-.CheckboxButton {
-  @include font-type-30;
-
-  cursor: pointer;
-  padding: 0.3125rem 0.625rem;
-  margin-right: 0.5rem;
-
-  display: flex;
-  align-items: center;
-
-  &--active {
-    background-color: $ux-blue-100;
-  }
-
-  &--bordered {
-    border: 1px solid $ux-gray-400;
-    border-radius: $ux-border-radius;
-  }
-}
