+++ src/CheckboxButtonGroup/index.js
@@ -1,7 +1,3 @@
 import CheckboxButtonGroup from './CheckboxButtonGroup';
-import CheckboxButton from './CheckboxButton';
 
-export {
-  CheckboxButtonGroup,
-  CheckboxButton,
-};
+export default CheckboxButtonGroup;
