+++ src/FormControlLabel/FormControlLabel.jsx
@@ -4,29 +4,44 @@ import classnames from 'classnames';
 
 import './FormControlLabel.scss';
 
-export default function FormControlLabel(props) {
+export default function FormControlLabel({
+  bordered,
+  checked,
+  className,
+  Control,
+  id,
+  text,
+  ...controlProps
+}) {
   return (
     <label
-      className={classnames('FormControlLabel', props.className)}
-      htmlFor={props.labelHtmlFor}
+      className={classnames(
+        'FormControlLabel',
+        className,
+        {
+          'FormControlLabel--bordered': bordered,
+          'FormControlLabel--active': bordered && checked,
+        },
+      )}
+      htmlFor={id}
     >
-      <span className="FormControlLabel__control">
-        {props.children}
-      </span>
-
-      {props.text}
+      <Control checked={checked} className="FormControlLabel__control" id={id} {...controlProps} />
+      {text}
     </label>
   );
 }
 
 FormControlLabel.propTypes = {
-  children: PropTypes.node.isRequired,
+  bordered: PropTypes.bool,
+  checked: PropTypes.bool,
   className: PropTypes.string,
-  labelHtmlFor: PropTypes.string,
+  Control: PropTypes.func.isRequired,
+  id: PropTypes.string.isRequired,
   text: PropTypes.string.isRequired,
 };
 
 FormControlLabel.defaultProps = {
+  bordered: false,
+  checked: undefined,
   className: undefined,
-  labelHtmlFor: '',
 };
