+++ src/FormControlLabel/FormControlLabel.scss
@@ -1,8 +1,25 @@
+@import '../../scss/theme';
+
 .FormControlLabel {
+  align-items: center;
   display: flex;
 
   &__control {
     display: flex;
     margin-right: .25rem;
+  }
+
+  &--bordered {
+    @include font-type-30;
+
+    border: 1px solid $ux-gray-400;
+    border-radius: $ux-border-radius;
+    cursor: pointer;
+    padding: 0.3125rem 0.625rem;
+    margin-right: 0.5rem;
+  }
+
+  &--active {
+    background-color: $ux-blue-100;
   }
 }
