+++ stories/Form Elements/FormControlLabel.stories.jsx
@@ -1,18 +1,23 @@
 import React from 'react';
+import { boolean, withKnobs } from '@storybook/addon-knobs';
 
-import { CheckboxButton } from 'src/CheckboxButtonGroup';
+import FormControlLabel from 'src/FormControlLabel';
+import CheckboxButton from 'src/CheckboxButton';
 
 export default {
   title: 'Design System/Form Elements/Form Control Label',
-  component: CheckboxButton,
+  component: FormControlLabel,
+  decorators: [withKnobs()],
 };
 
 export const Checkbox = () => (
-  <CheckboxButton
+  <FormControlLabel
+    bordered={boolean('Bordered button', false)}
+    Control={CheckboxButton}
     defaultValue="1"
     id="checkbox"
-    label="Labeled checkbox"
     name="checkbox"
+    text="Labeled checkbox"
     type="checkbox"
   />
 );
